move testing like code into test target
