+++ Wegli/Wegli/NumberPlateRecognizer/NumberPlateRecognizerService.swift
@@ -112,72 +112,3 @@ class NumberPlateRecognizerService {
         }
     }
 }
-
-extension NumberPlateRecognizerService {
-
-    // example scan bulk images
-    public func scanExampleImagesBulk(){
-        let exampleNumbers = ["32", "36", "39", "44", "52", "56", "79", "91"]
-        var exampleImagePaths = [String]()
-        var exampleImages = [UIImage]()
-
-        for number in exampleNumbers {
-            if let imagePath = Bundle.main.path(forResource: number, ofType: "jpg"), let image = UIImage(named: imagePath) {
-
-                if let last = imagePath.split(separator: "/").last {
-                    exampleImagePaths.append(last.description)
-                } else {
-                    exampleImagePaths.append(imagePath)
-                }
-
-                exampleImages.append(image)
-            } else {
-                print("Can not find image with \(number).jpg")
-            }
-        }
-
-        scan(images: exampleImages) { result in
-            switch result {
-                case .success(let plates):
-                    print(plates)
-                case .failure(let error):
-                    print(error)
-            }
-        }
-    }
-
-    // scan image after image and zip them together
-    public func scanExampleImages() {
-        let exampleNumbers = ["32", "36", "39", "44", "52", "56", "79", "91"]
-        var exampleImagePaths = [String]()
-        var exampleImages = [UIImage]()
-
-        for number in exampleNumbers {
-            if let imagePath = Bundle.main.path(forResource: number, ofType: "jpg"), let image = UIImage(named: imagePath) {
-
-                if let last = imagePath.split(separator: "/").last {
-                    exampleImagePaths.append(last.description)
-                } else {
-                    exampleImagePaths.append(imagePath)
-                }
-
-                exampleImages.append(image)
-            } else {
-                print("Can not find image with \(number).jpg")
-            }
-        }
-
-        zip(exampleImages, exampleImagePaths).forEach { (image, imagePath) in
-            scan(image: image, path: imagePath) { result in
-                switch result {
-                    case .failure(let error):
-                        print("WegLi - Recognizer: \(error)")
-                    case .success(let plates):
-                        for plate in plates {
-                            print("WegLi - Recognizer: \(plate)")
-                        }
-                }
-            }
-        }
-    }
-}
